sanders: Move AtFwd hal from device to framework manifest.
AtFwd hal is implemented as an android service and
it stars from system partition, that is, it's not
a traditional vendor hal and hence it should be
declared in the framework manifest and not in the
device manifest.
This commit introduces new device specific framework
manifest (framework_manifest.xml) and moves AtFwd hal
from device to framework manifest.

CRs-Fixed: 2287149
Change-Id: I897dc62ca2b9b04c3e7a0d3617c3c48d1c7ea4b0
Signed-off-by: Melvin <kartik3163@gmail.com>
Signed-off-by: melvinstark3 <iammelvin@pm.me>